requirements.txt
discord.py
python-dotenv
deepl
emoji-country-flag